Correction du FryDAO et du modèle.

+++ Tacos/app/dao/FryDAO.scala
@@ -16,7 +16,7 @@ trait FryComponent {
   class FryTable(tag: Tag) extends Table[Fry](tag, "frite") {
     def id = column[Long]("id", O.PrimaryKey, O.AutoInc) // Primary key, auto-incremented
     def name = column[String]("name")
-    def portion = column[Long]("portion")
+    def portion = column[Int]("portion")
     def price = column[Double]("price")
 
     // Map the attributes with the model.
